JAVA-12464 | fixing package names
