+++ core-kotlin-modules/core-kotlin-6/src/main/java/com/baeldung/accidentalOverride/Animal.java
@@ -1,0 +1,5 @@
+package com.baeldung.accidentalOverride;
+
+public interface Animal {
+    default String getSound() { return ""; }
+}
