--- core-kotlin-modules/core-kotlin-6/src/main/java/com/baeldung/accidental_override/Animal.java
@@ -1,5 +1,0 @@
-package com.baeldung.accidental_override;
-
-public interface Animal {
-    default String getSound() { return ""; }
-}
