+++ core-kotlin-modules/core-kotlin-6/src/main/kotlin/com/baeldung/accidentalOverride/Animals.kt
@@ -1,0 +1,25 @@
+package com.baeldung.accidentalOverride
+
+// This example will create an accidental override error, because of the conflicting `getSound()` function.
+//class Cat : Animal {
+//    val sound: String = "Meow"
+//}
+
+class CatWithCodeRemoved : Animal {
+// val sound: String = "Meow"
+}
+
+class CatWithValRenamed : Animal {
+    val myCatSound: String = "Meow"
+}
+
+class CatWithOverride : Animal {
+    override fun getSound(): String {
+        return "Meow"
+    }
+}
+
+class CatWithJvmName : Animal {
+    @get:JvmName("cat_sound")
+    val sound: String = "Meow"
+}
