+++ core-kotlin-modules/core-kotlin-advanced-2/src/main/java/com/baeldung/typeExposure/ImageSource.java
@@ -1,0 +1,10 @@
+package com.baeldung.typeExposure;
+
+public abstract class ImageSource {
+
+    public abstract MyImage produceImage();
+
+    static class MyImage {
+
+    }
+}
