+++ core-kotlin-modules/core-kotlin-advanced-2/src/main/java/com/baeldung/typeExposure/another/ImageConsumer.java
@@ -1,0 +1,10 @@
+package com.baeldung.typeExposure.another;
+
+import com.baeldung.typeExposure.ImageSource;
+
+public class ImageConsumer {
+
+    public void consumeImage(ImageSource is) {
+//        ImageSource.MyImage image = is.produceImage();
+    }
+}
