--- core-kotlin-modules/core-kotlin-advanced-2/src/main/java/com/baeldung/type_exposure/ImageSource.java
@@ -1,10 +1,0 @@
-package com.baeldung.type_exposure;
-
-public abstract class ImageSource {
-
-    public abstract MyImage produceImage();
-
-    static class MyImage {
-
-    }
-}
