--- core-kotlin-modules/core-kotlin-advanced-2/src/main/java/com/baeldung/type_exposure/another/ImageConsumer.java
@@ -1,10 +1,0 @@
-package com.baeldung.type_exposure.another;
-
-import com.baeldung.type_exposure.ImageSource;
-
-public class ImageConsumer {
-
-    public void consumeImage(ImageSource is) {
-//        ImageSource.MyImage image = is.produceImage();
-    }
-}
