+++ core-kotlin-modules/core-kotlin-advanced-2/src/main/kotlin/com/baeldung/typeExposure/RemoteImageSource.kt
@@ -1,0 +1,8 @@
+package com.baeldung.typeExposure
+
+internal class RemoteImageSource : ImageSource() {
+
+    override fun produceImage(): MyImage? {
+        return null
+    }
+}
