--- core-kotlin-modules/core-kotlin-advanced-2/src/main/kotlin/com/baeldung/type_exposure/RemoteImageSource.kt
@@ -1,8 +1,0 @@
-package com.baeldung.type_exposure
-
-internal class RemoteImageSource : ImageSource() {
-
-    override fun produceImage(): MyImage? {
-        return null
-    }
-}
