+++ core-kotlin-modules/core-kotlin-collections-3/src/main/kotlin/com/baeldung/iterateMap/Keys.kt
@@ -1,0 +1,18 @@
+package com.baeldung.iterateMap
+
+class Keys {
+    private val map = mapOf("Key1" to "Value1", "Key2" to "Value2", "Key3" to "Value3")
+    private fun keysForLoop() {
+        for (key in map.keys) {
+            val value = map[key]
+            print("$key : $value")
+        }
+    }
+
+    private fun keysForEachExtension() {
+        map.keys.forEach { key ->
+            val value = map[key]
+            print("$key : $value")
+        }
+    }
+}
