--- core-kotlin-modules/core-kotlin-collections-3/src/main/kotlin/com/baeldung/iterate_map/Values.kt
@@ -1,16 +1,0 @@
-package com.baeldung.iterate_map
-
-class Values {
-    private val map = mapOf("Key1" to "Value1", "Key2" to "Value2", "Key3" to "Value3")
-    private fun valuesForLoop() {
-        for (value in map.values) {
-            print(value)
-        }
-    }
-
-    private fun valuesForEachExtension() {
-        map.values.forEach { value ->
-            print(value)
-        }
-    }
-}
