+++ core-kotlin-modules/core-kotlin-lang-3/src/main/kotlin/com/baeldung/superNotExpression/SuperNotExpression.kt
@@ -1,0 +1,23 @@
+package com.baeldung.superNotExpression
+
+open class Vehicle() {
+    constructor(name: String) : this() {}
+
+    open fun start() {}
+}
+
+class Motorbike : Vehicle {
+    constructor(name: String) : super(name) {}
+
+    override fun start() {
+        super.start()
+    }
+}
+
+class Car() : Vehicle() {
+    init {
+        /*
+        super() //compilation error
+         */
+    }
+}
