--- core-kotlin-modules/core-kotlin-lang-3/src/main/kotlin/com/baeldung/super_not_expression/SuperNotExpression.kt
@@ -1,23 +1,0 @@
-package com.baeldung.super_not_expression
-
-open class Vehicle() {
-    constructor(name: String) : this() {}
-
-    open fun start() {}
-}
-
-class Motorbike : Vehicle {
-    constructor(name: String) : super(name) {}
-
-    override fun start() {
-        super.start()
-    }
-}
-
-class Car() : Vehicle() {
-    init {
-        /*
-        super() //compilation error
-         */
-    }
-}
