+++ core-kotlin-modules/core-kotlin-strings-2/src/main/kotlin/com/baeldung/charArrayToString/NumericStrings.kt
@@ -1,0 +1,18 @@
+package com.baeldung.charArrayToString
+
+object NumericStrings {
+    fun stringConstructor() {
+        val charArray = charArrayOf('b', 'a', 'e', 'l')
+        val convertedString = String(charArray)
+    }
+
+    fun joinToString() {
+        val charArray: Array<Char> = arrayOf('b', 'a', 'e', 'l')
+        val convertedString = charArray.joinToString()
+    }
+
+    fun stringBuilder() {
+        val charArray = charArrayOf('b', 'a', 'e', 'l')
+        val convertedString = StringBuilder().append(charArray).toString()
+    }
+}
