--- core-kotlin-modules/core-kotlin-strings-2/src/main/kotlin/com/baeldung/char_array_to_string/NumericStrings.kt
@@ -1,18 +1,0 @@
-package com.baeldung.char_array_to_string
-
-object NumericStrings {
-    fun stringConstructor() {
-        val charArray = charArrayOf('b', 'a', 'e', 'l')
-        val convertedString = String(charArray)
-    }
-
-    fun joinToString() {
-        val charArray: Array<Char> = arrayOf('b', 'a', 'e', 'l')
-        val convertedString = charArray.joinToString()
-    }
-
-    fun stringBuilder() {
-        val charArray = charArrayOf('b', 'a', 'e', 'l')
-        val convertedString = StringBuilder().append(charArray).toString()
-    }
-}
