+++ core-kotlin-modules/core-kotlin-strings-2/src/main/kotlin/com/baeldung/numericStrings/NumericStrings.kt
@@ -1,0 +1,22 @@
+package com.baeldung.numericStrings
+
+object NumericStrings {
+    object ToXTechnique {
+        fun isNumeric(toCheck: String): Boolean {
+            return toCheck.toDoubleOrNull() != null
+        }
+    }
+
+    object RegexTechnique {
+        fun isNumeric(toCheck: String): Boolean {
+            val regex = "-?[0-9]+(\\.[0-9]+)?".toRegex()
+            return toCheck.matches(regex)
+        }
+    }
+
+    object AllAndIsDigitTechnique {
+        fun isNumeric(toCheck: String): Boolean {
+            return toCheck.all { char -> char.isDigit() }
+        }
+    }
+}
