--- core-kotlin-modules/core-kotlin-strings-2/src/main/kotlin/com/baeldung/numeric_strings/NumericStrings.kt
@@ -1,22 +1,0 @@
-package com.baeldung.numeric_strings
-
-object NumericStrings {
-    object ToXTechnique {
-        fun isNumeric(toCheck: String): Boolean {
-            return toCheck.toDoubleOrNull() != null
-        }
-    }
-
-    object RegexTechnique {
-        fun isNumeric(toCheck: String): Boolean {
-            val regex = "-?[0-9]+(\\.[0-9]+)?".toRegex()
-            return toCheck.matches(regex)
-        }
-    }
-
-    object AllAndIsDigitTechnique {
-        fun isNumeric(toCheck: String): Boolean {
-            return toCheck.all { char -> char.isDigit() }
-        }
-    }
-}
