+++ core-kotlin-modules/core-kotlin-strings-2/src/test/kotlin/com/baeldung/numericStrings/NumericStringsUnitTest.kt
@@ -1,0 +1,19 @@
+package com.baeldung.numericStrings
+
+import com.baeldung.numericStrings.NumericStrings.ToXTechnique.isNumeric
+import org.junit.jupiter.api.Test
+import kotlin.test.assertTrue
+
+class NumericStringsUnitTest {
+    @Test
+    fun `check if number is string with toX methods`() {
+        assertTrue { isNumeric("11") }
+        assertTrue { isNumeric("-11") }
+        assertTrue { isNumeric("011") }
+        assertTrue { isNumeric("11.0F") }
+        assertTrue { isNumeric("11.0D") }
+        assertTrue { isNumeric("11.234") }
+        assertTrue { isNumeric("11.234e56") }
+        assertTrue { isNumeric("     123      ") }
+    }
+}
